Created a separate directory for parsing in src, added functions documentation
